Added Android Studio build files and app files to GitHub
These files represent the Fun Fact Activity application, which fetches
fun facts and displays them to users at the press of a button, while
changing the color of the button's text and the background of the screen
itself. This project was a valuable learning experience in understanding
views, layouts, and buttons in Android application programming.